Rename the dialog directives file.
